+++ finch/processes/wps_ensemble_indices_bbox.py
@@ -39,6 +39,7 @@ class XclimEnsembleBboxBase(FinchProcess):
             wpsio.start_date,
             wpsio.end_date,
             wpsio.ensemble_percentiles,
+            wpsio.average,
             wpsio.dataset_name,
             wpsio.copy_io(wpsio.rcp, min_occurs=1, max_occurs=3),
             wpsio.models,